+++ R/deprecated.R
@@ -81,6 +81,16 @@ setGeneric(
     }
 )
 
+#' @rdname deprecated
+#' @export
+setGeneric(
+    name = "plotReducedDim",
+    def = function(object, ...) {
+        .Deprecated("plotReduction")
+        standardGeneric("plotReducedDim")
+    }
+)
+
 
 
 ## nolint end